wp12_hackathon/dissemination_summary_prototype/prototype_b/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.2.1
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
durationpy==0.10
email_validator==2.2.0
fastapi==0.115.9
fastapi-cli==0.0.7
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
langchain==0.3.25
langchain-chroma==0.2.4
langchain-community==0.3.25
langchain-core==0.3.65
langchain-huggingface==0.3.0
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langsmith==0.3.45
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.5.0
mypy_extensions==1.1.0
narwhals==1.43.0
networkx==3.5
numpy==2.3.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.0
ollama==0.5.1
onnxruntime==1.22.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-asgi==0.55b1
opentelemetry-instrumentation-fastapi==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.3.0
pillow==11.2.1
posthog==5.0.0
propcache==0.3.2
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pypdf==5.6.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.45.3
streamlit==1.46.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
